Reorganize ui classes
