+++ app/src/androidTest/java/ch/epfl/sdp/drone3d/ConnectedDroneActivityTest.kt
@@ -10,6 +10,8 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import ch.epfl.sdp.drone3d.drone.DroneInstanceProvider.isConnected
+import ch.epfl.sdp.drone3d.ui.drone.ConnectedDroneActivity
+import ch.epfl.sdp.drone3d.ui.drone.DroneConnectActivity
 import junit.framework.Assert.assertEquals
 import org.junit.*
 import org.junit.runner.RunWith
