+++ app/src/androidTest/java/ch/epfl/sdp/drone3d/RegisterActivityTest.kt
@@ -12,6 +12,9 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.platform.app.InstrumentationRegistry
 import ch.epfl.sdp.drone3d.auth.AuthenticationModule
 import ch.epfl.sdp.drone3d.auth.AuthenticationService
+import ch.epfl.sdp.drone3d.ui.MainActivity
+import ch.epfl.sdp.drone3d.ui.auth.LoginActivity
+import ch.epfl.sdp.drone3d.ui.auth.RegisterActivity
 import com.google.android.gms.tasks.TaskCompletionSource
 import com.google.firebase.auth.AuthResult
 import dagger.hilt.android.testing.BindValue
