+++ app/src/main/res/layout/activity_drone_connect.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DroneConnectActivity">
+    tools:context=".ui.drone.DroneConnectActivity">
 
     <Button
         android:id="@+id/connect_button"
