+++ app/src/main/res/layout/activity_itinerary_create.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ItineraryCreateActivity">
+    tools:context=".ui.mission.ItineraryCreateActivity">
 
     <com.mapbox.mapboxsdk.maps.MapView
         android:id="@+id/mapView"
