+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/login_activity"
-    tools:context=".LoginActivity">
+    tools:context=".ui.auth.LoginActivity">
 
     <TextView
         android:id="@+id/emailText"
