+++ app/src/main/res/layout/activity_mapping_mission_selection.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MappingMissionSelectionActivity">
+    tools:context=".ui.mission.MappingMissionSelectionActivity">
 
     <ScrollView
         android:id="@+id/scrollView"
